Fix - ryanconfusion branch and queuesize of each vehicle

+++ petrolstation/src/com/company/vehicles/Vehicle.java
@@ -14,7 +14,7 @@ public abstract class Vehicle {
     public int fuelTankSize;
     public int fuelReceived;
     public double probabilityOfShop;
-    public int queueSize;
+    public double queueSize;
     public double queueDuration;
 
     /**
